Apply small fixes and extensions to base classes

+++ LilyPad/BDIM.pde
@@ -65,7 +65,7 @@ class BDIM{
     get_coeffs(body);
   }
   
-  BDIM( int n, int m, float dt, Body body, float nu, boolean QUICK, int u_inf){
+  BDIM( int n, int m, float dt, Body body, float nu, boolean QUICK, float u_inf){
     this(n,m,dt,body,new VectorField(n+2,m+2,u_inf,0),nu,QUICK);}
   BDIM( int n, int m, float dt, Body body, float nu, boolean QUICK ){
     this(n,m,dt,body,new VectorField(n+2,m+2,1,0),nu,QUICK);}
